pulls first name last name from parse
this update pulls the first name and last name of the residents and
populates the table with it. however does create a bug where on loop
around doesn’t repopulate table with name. just due to the new array
i’m populating the table with, should be able to fix when we’re
together. Also, I don’t think we need the loop back to table segue, the
ER docs said that the doctors only perform one evaluation after a shift
even though they have many residents, they just need to pick one.