attending and resident ID added to PFObject
All relevant Data is being uploaded. Only thing to fix now is whether
or not we want to loop and if we do we need to fix it since I changed
the array that is loading the table view. also there are a lot of
variables that we can clean up in this.